Time provider moved from state manager to util packag
