+++ packages/util/time_util/interface.go
@@ -1,0 +1,11 @@
+package time_util
+
+import (
+	"time"
+)
+
+type TimeProvider interface {
+	SetNow(time.Time)
+	GetNow() time.Time
+	After(time.Duration) <-chan time.Time
+}
