+++ packages/util/time_util/time_provider_artifficial_test.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-package sm_gpa_utils
+package time_util
 
 import (
 	"testing"
